Refactor how metadata about Pure is gathered and processed
This makes Pure's site depend on Pure via Bower, always. This allows the
website to gather and process metadata about Pure on startup, like the
version, filesizes, and current set of grid units.

Since gathering this metadata is done async, a new middleware function
has been added which will do all the async metadata gathering and
processing on the first request, _only_, after which the data is cached.