Remove defaults and add robust media query checks
This commit updates the `/start/` route so that it can
accept all the various routes. I updated the media query
parsing so that it is more robust in determining what is
and isnt a legitimate media query.

I also removed defaults, so if no values are specified in
the route, you just get an empty object sent to the client.