Setup dev-mode to create a Broccoli watcher and serve from Broccoli middleware
The watcher will keep the build/ dir updated to point to the current Broccoli tmp dir via a symlink, this ensures the rest of the app can seamlessly access files from the build dir.

Note: the Broccoli-related dependencies are npm `devDependencies`, therefore trying to serve from Broccoli in production will fail, on purpose.